Add HandPickup state
This works on the monitor from the Surface 3 and works on my MacBook
Pro, but the canvasView rectangle is way off when testing on the MacBook
Pro connected to the monitor. I suspect this is related to the Scale and
Layout settings (why it's different on the MacBook Pro vs. Surface 3 for
monitor). The canvas view is slightly off from the visible portion of
the canvas, I haven't figured out why.